actions/requirements.txt
mysql-connector-python==8.0.25
twilio==6.50.1
numpy==1.18.5
